requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0
selenium>=4.15.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
schedule>=1.2.0

# RAG Backend Dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
langchain>=1.0.0
langchain-google-genai>=2.0.0,<3.0.0
langchain-text-splitters>=0.3.0
chromadb>=0.4.22
langchain-community>=0.0.20
langchain-core>=1.0.0

